Fixed BUG 0010712 (attempt #2).
Committing fix from Thierry Carrard. Basically removed the time seeking
optimizations in vtkEnSightGolBinaryReader2 that are buggy. It solves the
freezing problem. For sure, time seeking optimization has to be put back at some
point, because sequential reading is way too slow on big files but the first
goal is to maintain functionality.

+++ Servers/Filters/vtkEnSightGoldBinaryReader2.h
@@ -185,7 +185,7 @@ protected:
 
   ifstream *IFile;
   // The size of the file could be used to choose byte order.
-  long FileSize;
+  int FileSize;
 
 private:
   vtkEnSightGoldBinaryReader2(const vtkEnSightGoldBinaryReader2&);  // Not implemented.
